+++ pyincore/globals.py
@@ -9,7 +9,7 @@ from logging import config as logging_config
 import logging
 import os
 
-PACKAGE_VERSION = "0.6.2"
+PACKAGE_VERSION = "0.6.3"
 
 INCORE_API_PROD_URL = "https://incore.ncsa.illinois.edu"
 TEST_INCORE_API_PROD_URL = "http://incore.ncsa.illinois.edu:31888"