+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - Pyincore enable local object of DFR3 Curve and Mapping [INCORE1-479](https://opensource.ncsa.illinois.edu/jira/browse/INCORE1-479)
+- Implement conditional fragility calculation and parametric [INCORE1-528](https://opensource.ncsa.illinois.edu/jira/browse/INCORE1-528)
 - User warning message when mapping fails due to mismatched datatype [INCORE1-559](https://opensource.ncsa.illinois.edu/jira/browse/INCORE1-559)
 - Methods to get uncertainty and variance for model based earthquakes [INCORE1-542](https://opensource.ncsa.illinois.edu/jira/browse/INCORE1-542)